refactor: created addPhrasesToTextarea. Closes #10
